+++ Cargo.toml
@@ -5,6 +5,4 @@ edition = "2021"
 
 [dependencies]
 alloy = { version = "0.1.2", features = ["rpc", "rpc-types-trace"] }
-
-
 serde = { version = "1.0", features = ["derive"] }
